+++ src/main/java/net/grid/vampiresdelight/integration/appleskin/AppleSkinEventHandler.java
@@ -1,7 +1,7 @@
 package net.grid.vampiresdelight.integration.appleskin;
 
 import de.teamlapen.vampirism.util.Helper;
-import net.grid.vampiresdelight.common.Configuration;
+import net.grid.vampiresdelight.common.VDConfiguration;
 import net.minecraft.client.Minecraft;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.api.distmarker.OnlyIn;
@@ -12,7 +12,7 @@ import squeek.appleskin.api.event.TooltipOverlayEvent;
 public class AppleSkinEventHandler {
     @SubscribeEvent
     public void onEvent(final TooltipOverlayEvent.Pre event) {
-        if (Configuration.SPECIAL_APPLE_SKIN_TOOLTIP.get() && Helper.isVampire(Minecraft.getInstance().player)) {
+        if (VDConfiguration.SPECIAL_APPLE_SKIN_TOOLTIP.get() && Helper.isVampire(Minecraft.getInstance().player)) {
             event.setCanceled(true);
         }
     }